chore: changing the name in README.md, too
Personally, I think the "o" should be lowercase because this is how it would look on the command line, as an imported module in Python, and as an installable package on PyPI (all of which have a lowercase convention), but it's your call. Capitalizing the "O" really makes you notice it, so there's that.

I'm pointing to this library on a [Gitter forum](https://app.gitter.im/#/room/#Scikit-HEP_uproot:gitter.im), so it's worthwhile being consistent about the name.